Begin refactoring MC sampling
